Themes: Move logged out layout setup to controller
Moved most /design specific handling out of client/boot/index.js so
that other logged out routes will work as normal.

Layout parameters are being passed to the layout component via
setState(), which could be causing the problems with React failing
to reconcile that component with the server-rendered version.
