+++ shared/my-sites/themes/controller.js
@@ -2,7 +2,8 @@
  * External Dependencies
  */
 var React = require( 'react' ),
-	titlecase = require( 'to-title-case' );
+	titlecase = require( 'to-title-case' ),
+	partialRight = require( 'lodash/function/partialRight' );
 
 /**
  * Internal Dependencies
@@ -54,7 +55,30 @@ var controller = {
 			} ),
 			document.getElementById( 'primary' )
 		);
+	},
+
+	themesLoggedOut: function( context ) {
+		const themesComponent = require( 'my-sites/themes/themes-selection' );
+		const getButtonOptions = require( 'my-sites/themes/theme-options' );
+		const themes = React.createFactory( themesComponent );
+
+		const themesProps = {
+			section: 'themes',
+			primary: themes( {
+				getOptions: partialRight( getButtonOptions, () => {}, () => {} ),
+				sites: false,
+				setSelectedTheme: () => {},
+				togglePreview: () => {},
+				secondary: null,
+				tertiary: null
+			})
+		};
+
+		// TODO: use a layout store instead of setting the state
+		context.layout.setState( themesProps );
 	}
 };
+
+
 
 module.exports = controller;
